requirements.txt
asgiref==3.2.10
Django==3.1
django-environ==0.4.5
gunicorn==20.0.4
numpy==1.19.1
pandas==1.1.1
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
whitenoise==5.2.0
